test(stella-observatory): split schema_conformance.rs by subject
crates/stella-observatory/tests/schema_conformance.rs had grown to exactly
1500 lines with no headroom for the next drift assertion (#4486). Split into
tests/schema_conformance/{route_survival,ratings_and_calls,recall_and_receipts,
diffs_and_agents,context_lifecycle}.rs by topic; the two real-migration
fixtures (real_store_workspace, real_context_workspace) and their seeding
helpers stay in the thin parent since every topic shares them.

This is a tests/*.rs integration-test crate root, not a src/-side module, so
its own submodules resolve beside it by default rather than under a
same-named subdirectory -- each mod declaration carries an explicit #[path]
to keep the split files inside tests/schema_conformance/ instead of littering
tests/ with files Cargo would otherwise auto-discover as their own separate
test binaries.

Every test name is unchanged, only relocated:
  route_survival.rs:
    every_route_survives_the_real_store_schema
    every_route_survives_a_store_older_than_this_build
  ratings_and_calls.rs:
    the_ratings_feed_counts_only_graded_turns_against_the_real_schema
    an_unfinished_execution_reports_its_tool_calls
    a_crashed_execution_recovers_its_calls_through_the_api
    an_abandoned_call_is_not_a_tool_error_on_the_leaderboard
  recall_and_receipts.rs:
    recall_latency_reaches_the_execution_detail
    a_recall_without_a_measurement_reads_as_null_not_zero
    sent_context_reconstructs_a_real_stores_receipt_and_verifies_it
  diffs_and_agents.rs:
    context_diff_names_the_moved_line_and_reports_identity_honestly
    session_turn_diff_replays_the_recorded_hunks
    the_agents_panel_separates_the_two_writers_against_the_real_schema
  context_lifecycle.rs:
    context_lifecycle_returns_the_promotion_lineage
    a_workspace_with_no_context_db_degrades_to_an_empty_lifecycle
    a_context_db_older_than_v8_degrades_to_empty_ledger_sections

No baseline entry added or needed post-split.

Closes #4486
