+++ crates/stella-observatory/tests/schema_conformance.rs
@@ -32,6 +32,18 @@
 //!    fresh workspace must render, not fail). Without this half, dropping a
 //!    table entirely would pass the no-500 check while silently blanking the
 //!    dashboard.
+//!
+//! # Where a test lives
+//!
+//! This file holds the two real-migration fixtures (`real_store_workspace`
+//! for `store.db`, `real_context_workspace` for `context.db`) and their
+//! seeding helpers — the part every topic needs — plus the `mod` declarations
+//! below. The `#[test]`s themselves live in `tests/schema_conformance/`,
+//! split by subject the same way #4486 found this file at the 1500-line
+//! ceiling with no headroom for the next drift assertion (#3923, #4479 had
+//! already had to relocate one witness elsewhere for exactly this reason). A
+//! new test joins the topic module it fits, or gets its own if none does; a
+//! new fixture stays here only if more than one topic needs it.
 
 use std::path::Path;
 
@@ -119,8 +131,9 @@ const ROUTES: &[(&str, Option<&str>)] = &[
     ),
     // The context.db half is empty in a store-only fixture — this workspace
     // deliberately does not build that file, and its real-schema gate is
-    // `context_lifecycle_returns_the_promotion_lineage` below (#1871). The
-    // pointer covers the half that reads `store.db`: the `parse_error` column
+    // `context_lifecycle_returns_the_promotion_lineage` in the
+    // `context_lifecycle` topic module (#1871). The pointer here covers the
+    // half that reads `store.db`: the `parse_error` column
     // added in schema v23 (#2175), which is exactly the kind of new column
     // this suite exists to prove resolvable.
     (
@@ -583,587 +596,6 @@ fn seed_receipt(store: &Store, execution_id: i64, call: &ToolCall, output: &Tool
         .expect("step manifest");
 }
 
-/// The gate itself: every route, against a real-migration database.
-#[test]
-fn every_route_survives_the_real_store_schema() {
-    let workspace = real_store_workspace();
-    let root: &Path = workspace.path();
-
-    for (route, seeded_pointer) in ROUTES {
-        let response = respond(root, route);
-        assert_eq!(
-            response.status,
-            "200 OK",
-            "{route} did not answer 200 — body: {}",
-            String::from_utf8_lossy(&response.body)
-        );
-        let body: serde_json::Value =
-            serde_json::from_slice(&response.body).unwrap_or_else(|e| panic!("{route}: {e}"));
-        assert!(
-            body.get("error").is_none(),
-            "{route} returned an error payload: {body}"
-        );
-        let Some(pointer) = seeded_pointer else {
-            continue;
-        };
-        assert!(
-            body.pointer(pointer).is_some(),
-            "{route} lost its seeded data at {pointer} — a column this crate reads \
-             was very likely renamed or dropped in stella-store. Body: {body}"
-        );
-    }
-}
-
-/// #2443, against the real schema: the ratings feed counts what the model
-/// graded, and the Improve tab's two panels agree about the same turn.
-///
-/// `crates/stella-observatory/src/db.rs`'s `HAS_SELF_REVIEW` is a SQL mirror of
-/// `stella_store::SelfReviewRow::is_empty` — mirrored because this crate must
-/// not link `stella-store` in production. This suite is what keeps the mirror
-/// honest: the store below is built through the real migration path by the real
-/// writers, so a column renamed underneath the predicate fails here.
-///
-/// The ids are read out of the payloads rather than assumed, which is also the
-/// product statement: the turn the lifecycle route names as unreadable is
-/// exactly the one the ratings feed must not present as an unrated reflection.
-#[test]
-fn the_ratings_feed_counts_only_graded_turns_against_the_real_schema() {
-    let workspace = real_store_workspace();
-    let root: &Path = workspace.path();
-
-    let lifecycle: serde_json::Value =
-        serde_json::from_slice(&respond(root, "/api/context-lifecycle").body).expect("json");
-    let unreadable: Vec<i64> = lifecycle["unreadable_reflections"]
-        .as_array()
-        .expect("unreadable_reflections")
-        .iter()
-        .map(|r| r["execution_id"].as_i64().expect("execution_id"))
-        .collect();
-    assert!(
-        !unreadable.is_empty(),
-        "the fixture seeds a parse failure through the real writer: {lifecycle}"
-    );
-
-    let reflections: serde_json::Value =
-        serde_json::from_slice(&respond(root, "/api/reflections").body).expect("json");
-    let ratings = reflections["ratings"].as_array().expect("ratings");
-    assert!(
-        !ratings.is_empty(),
-        "and a real self-review, which must still be listed: {reflections}"
-    );
-    for rating in ratings {
-        let id = rating["execution_id"].as_i64().expect("execution_id");
-        assert!(
-            !unreadable.contains(&id),
-            "execution {id} reflected unreadably and was never graded, so it is \
-             not a rating: {rating}"
-        );
-        let prose = |key: &str| rating[key].as_str().is_some_and(|s| !s.trim().is_empty());
-        assert!(
-            !rating["self_rating"].is_null()
-                || !rating["delivered"].is_null()
-                || prose("what_went_well")
-                || prose("what_to_improve")
-                || prose("critique"),
-            "a listed rating carries an assessment: {rating}"
-        );
-    }
-}
-
-/// The live projection, seen through the dashboard's own route.
-///
-/// This is the end-to-end statement of the v18 change: a turn that has made
-/// tool calls but has **not finished** reports them. Before v18 the
-/// projection was built by the turn finalizer, so this count was zero until
-/// the run ended — and stayed zero forever if it never did.
-#[test]
-fn an_unfinished_execution_reports_its_tool_calls() {
-    let dir = tempfile::TempDir::new().expect("tempdir");
-    let store = Store::open(dir.path()).expect("store");
-    let id = store
-        .begin_execution("run", "still going", "anthropic", "opus")
-        .expect("begin");
-    let call = ToolCall {
-        call_id: "c1".into(),
-        name: "read_file".into(),
-        input: serde_json::json!({ "path": "a.rs" }),
-    };
-    store
-        .record_event(id, 0, &AgentEvent::ToolStart { call: call.clone() })
-        .expect("start");
-    store
-        .record_event(
-            id,
-            1,
-            &AgentEvent::ToolResult {
-                call_id: "c1".into(),
-                output: ToolOutput::Ok {
-                    content: "fn a() {}".into(),
-                    data: None,
-                },
-                duration_ms: 5,
-                speculated: false,
-            },
-        )
-        .expect("result");
-    store
-        .record_event(
-            id,
-            2,
-            &AgentEvent::ToolStart {
-                call: ToolCall {
-                    call_id: "c2".into(),
-                    name: "bash".into(),
-                    input: serde_json::json!({ "command": "cargo test" }),
-                },
-            },
-        )
-        .expect("second start");
-    // No finish_execution: the turn is still running, which is exactly when
-    // an operator needs these numbers.
-
-    let executions: serde_json::Value =
-        serde_json::from_slice(&respond(dir.path(), "/api/executions").body).expect("json");
-    assert_eq!(
-        executions[0]["tool_calls"], 2,
-        "an in-flight turn reports its calls: {executions}"
-    );
-    assert!(
-        executions[0]["outcome"].is_null(),
-        "and is still visibly unfinished: {executions}"
-    );
-
-    let tools: serde_json::Value =
-        serde_json::from_slice(&respond(dir.path(), "/api/tools").body).expect("json");
-    let names: Vec<&str> = tools
-        .as_array()
-        .expect("array")
-        .iter()
-        .filter_map(|t| t["name"].as_str())
-        .collect();
-    assert!(
-        names.contains(&"read_file") && names.contains(&"bash"),
-        "the leaderboard sees both live calls: {names:?}"
-    );
-}
-
-/// A crashed run's calls come back from the log, through the dashboard.
-///
-/// The complement of the test above: the live write is what keeps the
-/// projection current, and the repair fold is what makes losing it
-/// survivable. Both have to be true at the route, not just in the store.
-#[test]
-fn a_crashed_execution_recovers_its_calls_through_the_api() {
-    let dir = tempfile::TempDir::new().expect("tempdir");
-    let store = Store::open(dir.path()).expect("store");
-    let id = store
-        .begin_execution("run", "died mid-turn", "anthropic", "opus")
-        .expect("begin");
-    for (seq, call_id) in ["c1", "c2", "c3"].iter().enumerate() {
-        store
-            .record_event(
-                id,
-                seq as u64,
-                &AgentEvent::ToolStart {
-                    call: ToolCall {
-                        call_id: (*call_id).into(),
-                        name: "bash".into(),
-                        input: serde_json::json!({ "command": "true" }),
-                    },
-                },
-            )
-            .expect("start");
-    }
-    // Exactly the state a pre-v18 crash left behind: every event on disk,
-    // and no projected row anywhere. Done with a raw connection rather than
-    // a store method — "delete the projection" is not an operation the
-    // production API should offer just so a test can stage a crash.
-    {
-        let raw =
-            rusqlite::Connection::open(dir.path().join(".stella/private/store.db")).expect("open");
-        raw.execute("DELETE FROM tool_calls WHERE execution_id = ?1", [id])
-            .expect("wipe the projection");
-    }
-    let before: serde_json::Value =
-        serde_json::from_slice(&respond(dir.path(), "/api/executions").body).expect("json");
-    assert_eq!(before[0]["tool_calls"], 0, "the calls are invisible");
-
-    let repaired = store.reconcile_interrupted_executions().expect("reconcile");
-
-    assert_eq!(repaired, 1);
-    let after: serde_json::Value =
-        serde_json::from_slice(&respond(dir.path(), "/api/executions").body).expect("json");
-    assert_eq!(
-        after[0]["tool_calls"], 3,
-        "every call is back, replayed from the log: {after}"
-    );
-}
-
-/// Context-recall latency reaches the execution detail (#875).
-///
-/// Recall is on the first-token path of every turn, so a slow one delays
-/// everything after it — and it was invisible in the dashboard, the receipt
-/// and the log alike. This drives the real event through the real store and
-/// reads it back through the real route.
-#[test]
-fn recall_latency_reaches_the_execution_detail() {
-    let dir = tempfile::TempDir::new().expect("tempdir");
-    let store = Store::open(dir.path()).expect("store");
-    let id = store
-        .begin_execution("run", "slow recall", "anthropic", "opus")
-        .expect("begin");
-    store
-        .record_event(
-            id,
-            0,
-            &AgentEvent::ContextRecall {
-                frames: vec![ContextFrameRef {
-                    id: Some("nod_1".into()),
-                    citation_label: "auth module".into(),
-                    provider: "workspace-memory".into(),
-                    source: "stella-context".into(),
-                    kind: "memory".into(),
-                    uri: None,
-                    method: None,
-                    token_cost: 90,
-                    block_id: None,
-                    content_digest: None,
-                }],
-                provider_mix: vec![ProviderShare {
-                    provider: "workspace-memory".into(),
-                    frames: 1,
-                }],
-                tokens: 90,
-                usage: None,
-                latency_ms: 1_450,
-                used_ann_index: Some(false),
-            },
-        )
-        .expect("recall event");
-
-    let detail: serde_json::Value =
-        serde_json::from_slice(&respond(dir.path(), &format!("/api/execution?id={id}")).body)
-            .expect("json");
-
-    assert_eq!(
-        detail["recall"][0]["latency_ms"], 1_450,
-        "the operator can see recall was the slow part: {detail}"
-    );
-    assert_eq!(
-        detail["recall"][0]["used_ann_index"], false,
-        "and that the accelerator did not fire, which is a different problem"
-    );
-    assert_eq!(detail["recall"][0]["frames"], 1);
-    assert_eq!(detail["recall"][0]["tokens"], 90);
-}
-
-/// A stream recorded before the field existed must read as "not measured",
-/// never as "instant" — `0 ms` would be a claim the data does not support.
-#[test]
-fn a_recall_without_a_measurement_reads_as_null_not_zero() {
-    let dir = tempfile::TempDir::new().expect("tempdir");
-    let store = Store::open(dir.path()).expect("store");
-    let id = store
-        .begin_execution("run", "legacy stream", "anthropic", "opus")
-        .expect("begin");
-    // Written straight to the table: this is a payload shaped the way a
-    // pre-#875 binary wrote it, which no current constructor can produce.
-    {
-        let raw =
-            rusqlite::Connection::open(dir.path().join(".stella/private/store.db")).expect("open");
-        raw.execute(
-            "INSERT INTO events (execution_id, seq, event_type, payload) \
-             VALUES (?1, 0, 'context_recall', ?2)",
-            rusqlite::params![
-                id,
-                r#"{"type":"context_recall","frames":[],"provider_mix":[],"tokens":0}"#
-            ],
-        )
-        .expect("legacy event");
-    }
-
-    let detail: serde_json::Value =
-        serde_json::from_slice(&respond(dir.path(), &format!("/api/execution?id={id}")).body)
-            .expect("json");
-
-    assert!(
-        detail["recall"][0]["latency_ms"].is_null(),
-        "unmeasured must not render as 0 ms: {detail}"
-    );
-    assert!(detail["recall"][0]["used_ann_index"].is_null());
-}
-
-/// Sent-context reconstruction, end to end against a real store (#1475).
-///
-/// The strongest form of the drift gate this suite exists for, because it
-/// checks a *byte-level* coupling rather than a column-level one. The
-/// observatory links no protocol crate, so it rebuilds a `tool_call` block's
-/// preimage by hand in `stella_protocol::ToolCall`'s declaration order; the
-/// receipt here was digested over that crate's own serializer. Reorder those
-/// fields and `verified` flips to false right here — the alternative being a
-/// silent "the journal is torn" banner on every dashboard in the field.
-#[test]
-fn sent_context_reconstructs_a_real_stores_receipt_and_verifies_it() {
-    let workspace = real_store_workspace();
-    let body = respond(
-        workspace.path(),
-        "/api/execution-context?id=1&turn=0&step=1&call_seq=0",
-    )
-    .body;
-    let v: serde_json::Value = serde_json::from_slice(&body).expect("json");
-    let context = &v["context"];
-    assert_eq!(context["found"], true, "{v}");
-    assert_eq!(
-        context["messages"][0]["body"], SYSTEM_PROMPT,
-        "the system prompt the model was actually sent: {v}"
-    );
-    assert_eq!(context["messages"][0]["role"], "system");
-    assert_eq!(
-        context["messages"][2]["body"], "← tool_result c1\nfn a() {}",
-        "the tool result, recovered from the journal: {v}"
-    );
-    assert_eq!(
-        context["verified"], true,
-        "a journal-resolved block stopped re-hashing to the digest \
-         stella_protocol's own serializer produced — a wire shape moved under \
-         the observatory's hand-built preimage: {v}"
-    );
-    // The era stamp is read from the real column the real writer set (#1981).
-    // This is the drift half of that signal: rename `executions.journal_era` in
-    // stella-store and the dashboard would quietly fall back to the legacy era
-    // for every execution, under-reporting a genuine integrity failure with no
-    // test saying a word. `tests/journal_era.rs` pins what the two eras mean;
-    // this pins that the column is still there to read.
-    assert_eq!(
-        context["journal_era"], "compaction_journaled",
-        "an execution this build's Store began must be stamped as journaling \
-         its compaction rewrites: {v}"
-    );
-    assert_eq!(context["digest_mismatch_severity"], "none", "{v}");
-}
-
-/// **The other half of the drift problem, and the one that actually bit.**
-///
-/// The gate above proves the queries work against a store migrated to *this*
-/// build's schema. But this crate opens every file `SQLITE_OPEN_READ_ONLY`
-/// and therefore never runs migrations — so it is routinely pointed at a
-/// store several versions *behind* the binary reading it. Upgrade `stella`,
-/// open the dashboard before running a single turn, and nothing has yet had
-/// any reason to open that file read-write.
-///
-/// Found by pointing this crate at an untouched copy of a real 54 MB store:
-/// every route referencing the v18 `tool_calls.state` column 500'd. A missing
-/// *table* had always degraded to an empty payload; a missing *column* did
-/// not, because `rusqlite` reports the two through different error variants
-/// (`SqliteFailure` vs `SqlInputError`) and only the first was matched.
-///
-/// So this drives every route against a database deliberately rolled back to
-/// the pre-v18 shape. None may error: an older store renders with a section
-/// empty, and fills in the moment a turn migrates it.
-#[test]
-fn every_route_survives_a_store_older_than_this_build() {
-    let workspace = real_store_workspace();
-    let db = workspace.path().join(".stella/private/store.db");
-    {
-        let raw = rusqlite::Connection::open(&db).expect("open");
-        // Rebuild `tool_calls` without `state` — exactly the shape a v17
-        // binary left behind. Dropping the column is the honest simulation;
-        // stubbing the queries would test the stub.
-        raw.execute_batch(
-            // The by-state index has to go first — it references the column,
-            // and SQLite refuses to leave an index pointing at nothing.
-            "DROP INDEX IF EXISTS tool_calls_by_state;
-             ALTER TABLE tool_calls DROP COLUMN state;
-             PRAGMA user_version = 17;",
-        )
-        .expect("roll the schema back");
-    }
-
-    for (route, _) in ROUTES {
-        let response = respond(workspace.path(), route);
-        assert_eq!(
-            response.status,
-            "200 OK",
-            "{route} failed against a pre-v18 store — a read-only observer never migrates, \
-             so it must degrade rather than 500. Body: {}",
-            String::from_utf8_lossy(&response.body)
-        );
-    }
-
-    // And the degradation is honest: the column is gone, so nothing is
-    // claimed to be running rather than a number being invented.
-    let cursor: serde_json::Value =
-        serde_json::from_slice(&respond(workspace.path(), "/api/v1/cursor").body).expect("json");
-    assert_eq!(cursor["tool_calls_running"], 0);
-    assert!(
-        cursor["events"].as_i64().unwrap_or(0) > 0,
-        "the columns that DO exist still report: {cursor}"
-    );
-}
-
-/// The #1511 witness: two calls of the same role whose system prefixes differ
-/// by one line produce exactly one hunk naming that line, and a
-/// byte-identical pair reports `changed: false`. Fails before the route
-/// exists (404), and fails if the baseline search stops at the execution
-/// boundary — the drift here is *across* turns, which is the only place a
-/// byte-stable system prompt can drift.
-#[test]
-fn context_diff_names_the_moved_line_and_reports_identity_honestly() {
-    let workspace = real_store_workspace();
-    let root: &Path = workspace.path();
-
-    // Across the turn boundary: execution 2's first worker call against
-    // execution 1's — the same role, one line of drift, system scope.
-    let body = respond(
-        root,
-        "/api/execution-context-diff?id=2&turn=0&step=1&call_seq=0&base=prev&only=system",
-    )
-    .body;
-    let diff: serde_json::Value = serde_json::from_slice(&body).expect("diff json");
-    assert_eq!(diff["found"], true, "{diff}");
-    assert_eq!(diff["changed"], true, "{diff}");
-    assert_eq!(diff["base"], "prev", "{diff}");
-    assert_eq!(diff["minimal"], true, "{diff}");
-    assert_eq!(diff["added"], 1, "one inserted line, one addition: {diff}");
-    let hunks = diff["hunks"].as_array().expect("hunks");
-    assert_eq!(hunks.len(), 1, "one contiguous change, one hunk: {diff}");
-    let added: Vec<&str> = hunks[0]["lines"]
-        .as_array()
-        .expect("lines")
-        .iter()
-        .filter(|l| l["op"] == "add")
-        .filter_map(|l| l["text"].as_str())
-        .collect();
-    assert_eq!(added, vec![DRIFT_LINE], "the diff names the moved line");
-    assert!(
-        diff["base_label"]
-            .as_str()
-            .unwrap_or_default()
-            .contains("execution 1"),
-        "a cross-turn baseline is unmistakable: {diff}"
-    );
-
-    // Inside the turn: step 2 against step 1, byte-identical by construction.
-    let body = respond(
-        root,
-        "/api/execution-context-diff?id=2&turn=0&step=2&call_seq=0&base=prev&only=system",
-    )
-    .body;
-    let same: serde_json::Value = serde_json::from_slice(&body).expect("diff json");
-    assert_eq!(
-        same["changed"], false,
-        "byte-identical is not a change: {same}"
-    );
-    assert_eq!(same["base"], "prev", "{same}");
-    assert!(
-        !same["base_label"]
-            .as_str()
-            .unwrap_or_default()
-            .contains("execution"),
-        "a same-turn baseline stays terse: {same}"
-    );
-
-    // A role's first call has no predecessor: the resolved base is reported
-    // as `prompt`, never silently claimed to be `prev`.
-    let body = respond(
-        root,
-        "/api/execution-context-diff?id=1&turn=0&step=1&call_seq=0&base=prev",
-    )
-    .body;
-    let first: serde_json::Value = serde_json::from_slice(&body).expect("diff json");
-    assert_eq!(
-        first["base"], "prompt",
-        "prev on a first call resolves: {first}"
-    );
-    assert_eq!(first["base_label"], "prompt as submitted", "{first}");
-}
-
-/// The #1870 observatory witness: a recorded turn diff replays through the
-/// route with its hunks naming the changed lines, the session view stamps
-/// the turn with its journal ordinal (the only persisted join between the
-/// two numbering schemes), and an unrecorded turn answers `found: false`
-/// with the full key set. Fails on main — the route is absent.
-#[test]
-fn session_turn_diff_replays_the_recorded_hunks() {
-    let workspace = real_store_workspace();
-    let root: &Path = workspace.path();
-
-    let body = respond(
-        root,
-        "/api/session-turn-diff?id=ses-1700000000000-424242&turn=1",
-    )
-    .body;
-    let v: serde_json::Value = serde_json::from_slice(&body).expect("json");
-    assert_eq!(v["found"], true, "{v}");
-    assert_eq!(v["execution_id"], 1, "the store-side turn identity: {v}");
-    assert_eq!(v["files"][0]["path"], "src/lib.rs", "{v}");
-    let lines = v["files"][0]["hunks"][0]["lines"]
-        .as_array()
-        .expect("lines");
-    let removed: Vec<&str> = lines
-        .iter()
-        .filter(|l| l["op"] == "remove")
-        .filter_map(|l| l["text"].as_str())
-        .collect();
-    let added: Vec<&str> = lines
-        .iter()
-        .filter(|l| l["op"] == "add")
-        .filter_map(|l| l["text"].as_str())
-        .collect();
-    assert_eq!(removed, vec!["two"], "the hunks name the old line: {v}");
-    assert_eq!(added, vec!["TWO"], "and the new one: {v}");
-
-    // The session view carries the join: the seeded execution's turn row is
-    // stamped with the journal ordinal its diff was recorded under.
-    let session: serde_json::Value =
-        serde_json::from_slice(&respond(root, "/api/session?id=ses-1700000000000-424242").body)
-            .expect("json");
-    let stamped = session["turns"]
-        .as_array()
-        .expect("turns")
-        .iter()
-        .find(|t| t["id"] == 1)
-        .expect("the seeded execution's turn row");
-    assert_eq!(
-        stamped["journal_turn"], 1,
-        "the page can only offer a diff through this stamp: {session}"
-    );
-
-    // Missing is a state: an unrecorded turn, and a store without the v21
-    // table at all, both answer the full (empty) shape.
-    let absent: serde_json::Value = serde_json::from_slice(
-        &respond(
-            root,
-            "/api/session-turn-diff?id=ses-1700000000000-424242&turn=99",
-        )
-        .body,
-    )
-    .expect("json");
-    assert_eq!(absent["found"], false, "{absent}");
-    assert_eq!(absent["files"], serde_json::json!([]), "{absent}");
-}
-
-/// The v30 `agent_uses.kind` column resolves against the real migrated schema,
-/// and the two writers arrive apart (#3822). The panel's grouping rule itself
-/// is witnessed in the crate's own tests; this is the schema-drift gate over
-/// it.
-#[test]
-fn the_agents_panel_separates_the_two_writers_against_the_real_schema() {
-    let workspace = real_store_workspace();
-    let session: serde_json::Value = serde_json::from_slice(
-        &respond(workspace.path(), "/api/session?id=ses-1700000000000-424242").body,
-    )
-    .expect("json");
-    let kinds: Vec<&str> = session["agents"]
-        .as_array()
-        .expect("agents panel")
-        .iter()
-        .filter_map(|a| a["kind"].as_str())
-        .collect();
-    assert!(kinds.contains(&"definition"), "{session}");
-    assert!(kinds.contains(&"delegation"), "{session}");
-}
-
 /// One ledger append through the real write API, with the record's own
 /// identity, hash and timestamps — the shape every production writer uses
 /// (`crates/stella-cli/src/memory/observations.rs::append_observation` et al).
@@ -1305,196 +737,24 @@ fn real_context_workspace() -> (tempfile::TempDir, String, String) {
     )
 }
 
-/// The #1871 witness: the route folds the seeded observation → proposal →
-/// promotion-event lineage back out of a real-migration `context.db`. Fails
-/// on main (the route is absent), and fails if any ledger or episode column
-/// this crate reads is renamed in `stella-context`.
-#[test]
-fn context_lifecycle_returns_the_promotion_lineage() {
-    let (workspace, proposal_lineage, observation_id) = real_context_workspace();
-
-    let body = respond(workspace.path(), "/api/context-lifecycle").body;
-    let v: serde_json::Value = serde_json::from_slice(&body).expect("json");
-    assert!(v.get("error").is_none(), "{v}");
-    assert_eq!(v["present"], true, "{v}");
-
-    let proposal = &v["proposals"][0];
-    assert_eq!(proposal["lineage_id"], proposal_lineage.as_str(), "{v}");
-    assert_eq!(proposal["candidate_id"], CANDIDATE_ID, "{v}");
-    assert_eq!(
-        proposal["status"], "confirmed",
-        "the decision standing is replayed from the event log, not stored: {v}"
-    );
-    assert_eq!(
-        proposal["supporting_observations"][0],
-        observation_id.as_str(),
-        "the lineage reaches back to its evidence: {v}"
-    );
-    assert_eq!(
-        proposal["events"][0]["action"], "confirmed",
-        "the proposal carries its own slice of the audit trail: {v}"
-    );
-
-    assert_eq!(v["events"][0]["action"], "confirmed", "{v}");
-    assert_eq!(
-        v["events"][0]["candidate_id"], CANDIDATE_ID,
-        "the timeline names the candidate its lineage points at: {v}"
-    );
-
-    assert_eq!(v["episodes"][0]["outcome"], "success", "{v}");
-    assert_eq!(v["episodes"][0]["summary"], "added the parser fix", "{v}");
-
-    let kinds: Vec<&str> = v["counts"]
-        .as_array()
-        .expect("counts")
-        .iter()
-        .filter_map(|c| c["kind"].as_str())
-        .collect();
-    for kind in ["observation", "record_proposal", "promotion_event"] {
-        assert!(kinds.contains(&kind), "counts missing {kind}: {v}");
-    }
-}
-
-/// Missing is a state: a workspace that has never built a context plane
-/// answers with the full (empty) payload shape, never a 500 and never a
-/// missing key.
-#[test]
-fn a_workspace_with_no_context_db_degrades_to_an_empty_lifecycle() {
-    let workspace = real_store_workspace();
-    let response = respond(workspace.path(), "/api/context-lifecycle");
-    assert_eq!(
-        response.status,
-        "200 OK",
-        "body: {}",
-        String::from_utf8_lossy(&response.body)
-    );
-    let v: serde_json::Value = serde_json::from_slice(&response.body).expect("json");
-    assert_eq!(v["present"], false, "{v}");
-    for key in [
-        "counts",
-        "proposals",
-        "events",
-        "episodes",
-        "selection_health",
-    ] {
-        assert_eq!(v[key], serde_json::json!([]), "{key} must be empty: {v}");
-    }
-}
-
-/// The read-only observer never migrates, so it can be pointed at a
-/// `context.db` older than the v8 lifecycle ledger — same hazard the pre-v18
-/// store test above covers. The ledger sections degrade to empty and the
-/// episode list (whose v8 columns are also gone) degrades with them; nothing
-/// 500s, and everything fills in after the next session migrates the file.
-#[test]
-fn a_context_db_older_than_v8_degrades_to_empty_ledger_sections() {
-    let (workspace, _, _) = real_context_workspace();
-    {
-        let raw = rusqlite::Connection::open(workspace.path().join(".stella/private/context.db"))
-            .expect("open");
-        // Rebuild the pre-v8 shape honestly: no ledger table, no lineage
-        // columns on `episode`. Dropping the whole table also drops its
-        // append-only triggers, exactly as a pre-v8 file never had them.
-        raw.execute_batch(
-            "DROP TABLE context_records;
-             DROP INDEX IF EXISTS idx_episode_lineage;
-             ALTER TABLE episode DROP COLUMN lineage_id;
-             ALTER TABLE episode DROP COLUMN superseded_at;
-             PRAGMA user_version = 7;",
-        )
-        .expect("roll the schema back");
-    }
-
-    let response = respond(workspace.path(), "/api/context-lifecycle");
-    assert_eq!(
-        response.status,
-        "200 OK",
-        "a pre-v8 context.db must degrade, not 500 — body: {}",
-        String::from_utf8_lossy(&response.body)
-    );
-    let v: serde_json::Value = serde_json::from_slice(&response.body).expect("json");
-    assert_eq!(v["present"], true, "the file exists and is reported: {v}");
-    for key in [
-        "counts",
-        "proposals",
-        "events",
-        "episodes",
-        "selection_health",
-    ] {
-        assert_eq!(v[key], serde_json::json!([]), "{key} must be empty: {v}");
-    }
-}
-
-/// An interrupted turn's in-flight call must not count against its tool.
-///
-/// The sweep settles a still-`running` call as `'abandoned'` (v24), and the
-/// leaderboard reports it in its own column: before #3146 it landed in
-/// `errors`, so every ctrl-C or SIGKILL charged an "error" to whatever tool
-/// happened to be in flight — usually the slowest, most-watched ones.
-#[test]
-fn an_abandoned_call_is_not_a_tool_error_on_the_leaderboard() {
-    let dir = tempfile::TempDir::new().expect("tempdir");
-    let store = Store::open(dir.path()).expect("store");
-    let id = store
-        .begin_execution("run", "interrupted mid-tool", "anthropic", "opus")
-        .expect("begin");
-    store
-        .record_event(
-            id,
-            0,
-            &AgentEvent::ToolStart {
-                call: ToolCall {
-                    call_id: "c1".into(),
-                    name: "bash".into(),
-                    input: serde_json::json!({ "command": "false" }),
-                },
-            },
-        )
-        .expect("start c1");
-    store
-        .record_event(
-            id,
-            1,
-            &AgentEvent::ToolResult {
-                call_id: "c1".into(),
-                output: ToolOutput::error("exit status 1"),
-                duration_ms: 3,
-                speculated: false,
-            },
-        )
-        .expect("c1 fails for real");
-    store
-        .record_event(
-            id,
-            2,
-            &AgentEvent::ToolStart {
-                call: ToolCall {
-                    call_id: "c2".into(),
-                    name: "bash".into(),
-                    input: serde_json::json!({ "command": "sleep 60" }),
-                },
-            },
-        )
-        .expect("start c2");
-    // The owner is gone: the sweep settles c2 as abandoned, not errored.
-    store.mark_execution_interrupted(id).expect("sweep");
-
-    let tools: serde_json::Value =
-        serde_json::from_slice(&respond(dir.path(), "/api/tools").body).expect("json");
-    let bash = tools
-        .as_array()
-        .expect("array")
-        .iter()
-        .find(|t| t["name"] == "bash")
-        .cloned()
-        .expect("bash row");
-    assert_eq!(bash["calls"], 2, "both calls count: {bash}");
-    assert_eq!(
-        bash["errors"], 1,
-        "only the real failure is an error (#3146): {bash}"
-    );
-    assert_eq!(
-        bash["abandoned"], 1,
-        "the abandonment is visible, apart from the error count: {bash}"
-    );
-}
+// ── Topic modules ──────────────────────────────────────────────────────
+// Each declares `use super::*;` to reach the fixtures and imports above.
+// New tests join the topic they fit, or start a new module if none does.
+//
+// This file is a `tests/*.rs` integration-test crate ROOT, so its own
+// submodules resolve beside it (`tests/<name>.rs`) by default -- unlike a
+// `src/`-side split, where a submodule of `foo.rs` resolves under `foo/`.
+// The explicit `#[path]` on each is what keeps the topic files inside
+// `tests/schema_conformance/` instead of littering `tests/` itself with
+// files Cargo would otherwise try to auto-discover as their own separate
+// integration-test binaries.
+#[path = "schema_conformance/context_lifecycle.rs"]
+mod context_lifecycle;
+#[path = "schema_conformance/diffs_and_agents.rs"]
+mod diffs_and_agents;
+#[path = "schema_conformance/ratings_and_calls.rs"]
+mod ratings_and_calls;
+#[path = "schema_conformance/recall_and_receipts.rs"]
+mod recall_and_receipts;
+#[path = "schema_conformance/route_survival.rs"]
+mod route_survival;
